Towards shared modules: refactor.
